Revise project description in README.md
My project is called My World In Data which website that allows the user to choose the color of the background image of the dashboard OR upload their own image. Used as a to-do list, notes, and has a motivational phrase generator. I made it using html, css, and js. It has a default background of gradient green, but users can change it by either choosing a color from the box at the bottom and clicking a button or uploading their own photo. I originally wanted to create another page that links to the dashboard, but it didn't work, so I put all my features onto one page. Then, I created a to-do list, followed by the notes and motivational quote. I had to make sure it had a scroll feature, fits perfectly with all screens (resizes), and to not overflow. I worked on this project simaltaniously and by sections. So first I did the to-do list with the html, css, and javascript; moved onto notes, etc. I used tutorials for the to-do list, and both the background image and the workings were after trying to do it on my own. I encountered several issues throughout, particularly with the background working and replacing the default one. I also ran into issues with the motivational phrases. With this, I learnt how to make these features in a project, but also how everything interwines with one another and how even the smallest error or issues can complicate a project by so much (something I fell for A LOT). This also did teach me patience though and to not get frustrated and mainly, always be open to trying new methods in order to get your code to work.